fix tu add typescript hmi

+++ unittests/mock/cmp/MockAds1115.h
@@ -11,7 +11,7 @@ namespace Component
 			MOCK_METHOD0(Initialize, Core::Status( void ));
 			MOCK_METHOD1(Update, void( const uint64_t ));
 
-			MOCK_CONST_METHOD1(ReadADC, int16_t( const Ads1115Pin ));
+			MOCK_CONST_METHOD1(ReadADC, float( const Ads1115Pin ));
 		};
 	}
 }
